+++ fabric/src/client/kotlin/miragefairy2024/client/fabric/MirageFairy2024FabricReiClientPlugin.kt
@@ -7,22 +7,23 @@ import me.shedaniel.rei.api.client.registry.screen.ScreenRegistry
 import me.shedaniel.rei.api.common.entry.comparison.ItemComparatorRegistry
 import miragefairy2024.client.mod.recipeviewer.rei.ReiClientEvents
 import miragefairy2024.mod.recipeviewer.rei.ReiEvents
+import mirrg.kotlin.helium.Single
 
 @Suppress("unused")
 class MirageFairy2024FabricReiClientPlugin : REIClientPlugin {
     override fun registerCategories(registry: CategoryRegistry) {
-        ReiClientEvents.onRegisterCategories.fire { it(registry) }
+        ReiClientEvents.onRegisterCategories.fire { it(Single(registry)) }
     }
 
     override fun registerDisplays(registry: DisplayRegistry) {
-        ReiClientEvents.onRegisterDisplays.fire { it(registry) }
+        ReiClientEvents.onRegisterDisplays.fire { it(Single(registry)) }
     }
 
     override fun registerItemComparators(registry: ItemComparatorRegistry) {
-        ReiEvents.onRegisterItemComparators.fire { it(registry) }
+        ReiEvents.onRegisterItemComparators.fire { it(Single(registry)) }
     }
 
     override fun registerScreens(registry: ScreenRegistry) {
-        ReiClientEvents.onRegisterScreens.fire { it(registry) }
+        ReiClientEvents.onRegisterScreens.fire { it(Single(registry)) }
     }
 }
