+++ neoforge/src/main/kotlin/miragefairy2024/client/neoforge/MirageFairy2024NeoForgeReiClientPlugin.kt
@@ -8,23 +8,24 @@ import me.shedaniel.rei.api.common.entry.comparison.ItemComparatorRegistry
 import me.shedaniel.rei.forge.REIPluginClient
 import miragefairy2024.client.mod.recipeviewer.rei.ReiClientEvents
 import miragefairy2024.mod.recipeviewer.rei.ReiEvents
+import mirrg.kotlin.helium.Single
 
 @REIPluginClient
 @Suppress("unused")
 class MirageFairy2024NeoForgeReiClientPlugin : REIClientPlugin {
     override fun registerCategories(registry: CategoryRegistry) {
-        ReiClientEvents.onRegisterCategories.fire { it(registry) }
+        ReiClientEvents.onRegisterCategories.fire { it(Single(registry)) }
     }
 
     override fun registerDisplays(registry: DisplayRegistry) {
-        ReiClientEvents.onRegisterDisplays.fire { it(registry) }
+        ReiClientEvents.onRegisterDisplays.fire { it(Single(registry)) }
     }
 
     override fun registerItemComparators(registry: ItemComparatorRegistry) {
-        ReiEvents.onRegisterItemComparators.fire { it(registry) }
+        ReiEvents.onRegisterItemComparators.fire { it(Single(registry)) }
     }
 
     override fun registerScreens(registry: ScreenRegistry) {
-        ReiClientEvents.onRegisterScreens.fire { it(registry) }
+        ReiClientEvents.onRegisterScreens.fire { it(Single(registry)) }
     }
 }
